Move `move` from `entry`
