+++ include/tmp/file
@@ -170,6 +170,20 @@ public:
   /// @returns An opened stream for output operations
   std::ofstream output_stream(std::ios::openmode mode = std::ios::trunc) const;
 
+  /// Moves the managed file to a given target, releasing
+  /// ownership of the managed file; behaves like `std::filesystem::rename`
+  /// even when moving between filesystems
+  /// @param[in] to A path to the target file
+  /// @throws std::filesystem::filesystem_error if cannot move the owned file
+  void move(const std::filesystem::path& to);
+
+  /// Moves the managed file recursively to a given target, releasing
+  /// ownership of the managed file; behaves like `std::filesystem::rename`
+  /// even when moving between filesystems
+  /// @param[in]  to A path to the target file
+  /// @param[out] ec Parameter for error reporting
+  void move(const std::filesystem::path& to, std::error_code& ec);
+
   /// Deletes the managed file
   ~file() noexcept override;
