+++ src/file.cpp
@@ -2,6 +2,7 @@
 #include <tmp/file>
 
 #include "create.hpp"
+#include "move.hpp"
 
 #include <array>
 #include <cstddef>
@@ -347,6 +348,22 @@ std::ifstream file::input_stream() const {
 std::ofstream file::output_stream(std::ios::openmode mode) const {
   binary ? mode |= std::ios::binary : mode &= ~std::ios::binary;
   return std::ofstream(path(), mode);
+}
+
+void file::move(const fs::path& to) {
+  std::error_code ec;
+  move(to, ec);
+
+  if (ec) {
+    throw fs::filesystem_error("Cannot move a temporary file", path(), to, ec);
+  }
+}
+
+void file::move(const fs::path& to, std::error_code& ec) {
+  tmp::move(*this, to, ec);
+  if (!ec) {
+    clear();
+  }
 }
 
 file::~file() noexcept {
